cheap_RSA/requirements.txt
Flask==3.0.0
pycryptodome==3.20.0
sympy==1.12